CMakeLists: Adjust for CMake 4.0's removal of deprecated syntax.
CMake 4.0 is more strict than previous versions, and has enforced
removals of deprecated syntax. Compatibility with versions of CMake
prior to 3.5 is no longer permitted either, so we'll change it to say
4.0 to fix that problem. After that, the custom commands for generating
the manual page need to be adjusted as CMake now needs to know when to
run commands, so we will generate the manual page after the package is
built by passing POST_BUILD to the custom commands.

Tested with CMake 4.0 and 3.31.6.